modify

+++ TBBStockView/Source/TBBTimeLine/TBBTimeLineBottomView/TBBTimeLineBottomView.m
@@ -127,7 +127,7 @@
                 }else if (lineModel.volume > ws.maxVolume) {
                     ws.maxVolume = lineModel.volume;
                 }else if (lineModel.volume < 0){
-                    NSLog(@"第%ld个成交量%.2f小于0",idx,lineModel.volume);
+                    NSLog(@"第%ld个成交量%.2f小于0", (unsigned long)idx,lineModel.volume);
                 }
             }];
             NSLog(@"minVolume: %.2f---maxVolume: %.2f",self.minVolume,self.maxVolume);
